third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-stretch-expected.txt

This is a testharness.js-based test.
[PASS] Can set 'font-stretch' to CSS-wide keywords: initial
[PASS] Can set 'font-stretch' to CSS-wide keywords: inherit
[PASS] Can set 'font-stretch' to CSS-wide keywords: unset
[PASS] Can set 'font-stretch' to CSS-wide keywords: revert
[PASS] Can set 'font-stretch' to var() references:  var(--A)
[FAIL] Can set 'font-stretch' to the 'normal' keyword: normal
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'ultra-condensed' keyword: ultra-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'extra-condensed' keyword: extra-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'condensed' keyword: condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'semi-condensed' keyword: semi-condensed
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'semi-expanded' keyword: semi-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'expanded' keyword: expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'extra-expanded' keyword: extra-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[FAIL] Can set 'font-stretch' to the 'ultra-expanded' keyword: ultra-expanded
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
[PASS] Can set 'font-stretch' to a percent: 0%
[FAIL] Can set 'font-stretch' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3
[FAIL] Can set 'font-stretch' to a percent: 3.14%
  assert_approx_equals: expected 3.14 +/- 0.000001 but got 3
[PASS] Can set 'font-stretch' to a percent: calc(0% + 0%)
[PASS] Setting 'font-stretch' to a length: 0px throws TypeError
[PASS] Setting 'font-stretch' to a length: -3.14em throws TypeError
[PASS] Setting 'font-stretch' to a length: 3.14cm throws TypeError
[PASS] Setting 'font-stretch' to a length: calc(0px + 0em) throws TypeError
[PASS] Setting 'font-stretch' to a time: 0s throws TypeError
[PASS] Setting 'font-stretch' to a time: -3.14ms throws TypeError
[PASS] Setting 'font-stretch' to a time: 3.14s throws TypeError
[PASS] Setting 'font-stretch' to a time: calc(0s + 0ms) throws TypeError
[PASS] Setting 'font-stretch' to an angle: 0deg throws TypeError
[PASS] Setting 'font-stretch' to an angle: 3.14rad throws TypeError
[PASS] Setting 'font-stretch' to an angle: -3.14deg throws TypeError
[PASS] Setting 'font-stretch' to an angle: calc(0rad + 0deg) throws TypeError
[PASS] Setting 'font-stretch' to a flexible length: 0fr throws TypeError
[PASS] Setting 'font-stretch' to a flexible length: 1fr throws TypeError
[PASS] Setting 'font-stretch' to a flexible length: -3.14fr throws TypeError
[PASS] Setting 'font-stretch' to a number: 0 throws TypeError
[PASS] Setting 'font-stretch' to a number: -3.14 throws TypeError
[PASS] Setting 'font-stretch' to a number: 3.14 throws TypeError
[PASS] Setting 'font-stretch' to a number: calc(2 + 3) throws TypeError
[PASS] Setting 'font-stretch' to a transform: translate(50%, 50%) throws TypeError
[PASS] Setting 'font-stretch' to a transform: perspective(10em) throws TypeError
[PASS] Setting 'font-stretch' to a transform: translate3d(0px, 1px, 2px) translate(0px, 1px) rotate3d(1, 2, 3, 45deg) rotate(45deg) scale3d(1, 2, 3) scale(1, 2) skew(1deg, 1deg) skewX(1deg) skewY(45deg) perspective(1px) matrix3d(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16) matrix(1, 2, 3, 4, 5, 6) throws TypeError
Harness: the test ran to completion.
